Make it faster to start a new shell
Change various scripts to avoid running commands as part of starting a
new shell, such as `brew --prefix`, as the outputs of these commands
don't change frequently enough. Also make sure the zcompdump file is
only replaced when needed - probably when FPATH is changed.

Also revert the change to the tests to make the macOS compatible as they
were broken in Docker.